back/src/gen-user/requirements.txt
qrcode[pil]
pyotp
psycopg2-binary
cryptography
python-dotenv
